Implemented fix for when column is enclosed in parens (round brackets) and single byte as well as multibyte table and column aliases.  (#1812)
This pull request fix the issue when column name is enclosed in parens (round brackets) and no alias is specified.
It also fix the implementation for column and table aliases contain both single byte and multibyte characters whose length less than or more than 64 bytes.

*Issue 1*: Handling of column name when enclosed by any number of parens. When column is enclosed in parens
(round brackets) and no alias is specified, then the column name in result set is returned incorrectly. In such cases
the actual column name should come, instead from first characters of query till length of column name. To resolve
this issue I added a condition in pre_transform_target_entry that will handle multiple levels of parens and spaces in
between parens.

*Issue 2* : Handle column and table aliases. Column and table alias does not work when alias length is more than or
equals to 64 bytes in case of multibyte aliases. When query select 1 as '您对“数据一览“中的车型，颜色，内饰，选装, ';
executes through T-SQL endpoint, it causes client to hang and throws protocol error in TDS stream. The logic of
pre_transform_target_entry was not handling the aliases contain multibyte characters whose length is more than 64
bytes accurately and overwriting the alias name in case of truncation. To address this problem, I redefined the
function and implemented the conditions which will extract identifier name using extract_identifier() and handle both
the case when aliases are not delimited as well as delimited with square bracket, single quotes and double quotes.

*Issue 3* : Aliases with multibyte chars cause MVU failure. When establishing a view of alias with multibyte chars
through T-SQL endpoint and retrieving it's definition from PG end point, extra unreadable chars are appearing at the
end of the alias and even we can not use this view through tsql endpoint as it throws protocol error in TDS stream.
But while we use similar alias on PG end point, we are able to retrieve it's correct definition. This discrepancy in
behaviour is causing MVU failure. After implementing the fix, the correct alias definition is obtained from the PG
endpoint, which rectifies the issue.

Task: BABEL-4384, BABEL-4231
Signed-off-by: Riya Jain <riyaajn@amazon.com>