Fix the issue of preserving case of column names in SELECT corrname.colname. (#1866)
This commit fixes preserving the original case spelling of column names while selecting named column with a
correlation name.

Issue: When selecting named column with a correlation name, the lowercase name is always returned. The case of
column name is not preserved. For instance, select ABC from table, here ABC is the column name. In this case column
name becomes downcased. To address this issue, I've made some code modifications to ensure that the column name
is handled correctly, whether it's truncated or its case needs to be preserved.

Task: BABEL-4279
Signed-off-by: Riya Jain <riyaajn@amazon.com>