Fix for the crash in pre_transform_target_entry (#1931)
This pull request fix the crash is `pre_transform_target_entry` for queries that involves DELETE ... OUTPUT with JOIN statement. 

For the delete queries with join statement, the `res->location` will be passed zero. This causes server to crash as the value of `last_dot` pointer has never been computed. This commit implements a check which first computed whether the `res->location` is correct or not and then further proceed to find original identifier. 

Task: BABEL-4484
Signed-off-by: Riya Jain <riyaajn@amazon.com>