fix typmod value in case of nvarchar(max), varchar(max), varbinary(max) (#2036)
This PR fixes a crash in TdsSendTypeNVarchar for nvarchar(max) by resolving a bug in handling varchar(max) , nvarchar(max) and varbinary(max) and limits a maximum scale value for (var)char, (var)binary, and (n)(var)char types. To do that, we set a limit of 8000 for (var)char and (var)binary, and a limit of 4000 for (n)(var)char.

Signed-off-by: Sandeep Kumawat skumwt@amazon.com
Issues Resolved

BABEL-4547