Fix issues with TVFs which use identifiers longer than 63 characters (#2430) (#2558)
In Table-valued functions which return data into a table variable, we create a temporary table in which to return the data. The name of this temporary table is generated as a combination of the table variable and function name. If this combination exceeds 63 characters (the limit of identifiers in postgresql), then we can run into issues trying to refer back to the temporary table to insert the data while executing the function, as the name used to access it will be different than what is stored in the postgres catalog due to truncation.

Fix this by utilizing the pre-existing downcase_truncate_identifier() function to ensure that the new identifier is within the postgresql character limit. Additionally, truncate the temporary table name (which normally tries to append the @@nest_level to the end) to 63 characters to prevent issues there.

Task: BABEL-3311

Signed-off-by: Jason Teng <jasonten@amazon.com>