Fix crash when trying to delete a trigger on a temp table (#2559)
Triggers do not have visibility into temp tables from other sessions. In particular, if a trigger is created on a temp table, and then a separate session attempts to drop that trigger, the reference temp table will not be visible within the second session, and a null OID reference will be passed through the internal code. Currently this causes a segmentation fault, so fix it by detecting for this case and erroring out instead.

Task: BABEL-4867

Signed-off-by: Jason Teng <jasonten@amazon.com>