Fix differences in T-SQL COALESCE function while calling the function with variables and constants together (#2635)
This commit will fix the following issues with T-SQL COALESCE function :

When the arguments contain the mix of variables and constants as shown below, the below query should return 4 as it is the first non-null value. But BBF throws error as we try to evaluate the constant ('dfidhfgi') at the compile time in this case.

DECLARE @abc int = 4 ;
select COALESCE(@abc, 'dfidhfgi')

When the arguments are of varbinary and nvarchar type, BBF throws error as we have defined the CAST between varbinary and nvarchar as ASSIGNMENT, but in this case IMPLICIT casting is used.

select coalesce(CAST('x'AS VARBINARY), CAST('x' AS NVARCHAR(4000)), 'x')

Issues Resolved: BABEL-726
Engine PR : babelfish-for-postgresql/postgresql_modified_for_babelfish#380

Signed-off-by: Sai Rohan Basa bsrohan@amazon.com