Added wrapper functions for REPLACE() Function to fix issue with input containing multi-byte characters. (#2790)
Currently for string function REPLACE(), the existing definition takes input as type TEXT and return type is TEXT. This
leads to incorrect output for input containing multi-byte characters with NCHAR and NVARCHAR datatype for input, as
the return datatype should have been NVARCHAR for NCHAR and NVARCHAR datatype.

This commit will fix this issue by adding wrapper functions with correct input and return datatype for REPLACE()
function.

This commit also updates the castcontext of CAST from bbf_binary to varchar, to IMPLICIT which was previously
marked as ASSIGNMENT. Due to this REPLACE function call with input type binary was not able choose the available
replace definitions with argument datatype VARCHAR/NVARCHAR.

Task: BABEL-4836
Signed-off-by: Rohit Bhagat <rohitbgt@amazon.com>