Fix the index creation error on table created through SELECT INTO  (#2893)
If a table is formed via SELECT INTO command then the column name does not get converted to lowercase.
Due to this, indexation over that column fails which is the issue we want to resolve.
This happens as the SELECT INTO command does not convert the column name in lowercase(save it as it is) whereas the index creation command search for the column name(over which indexation is done) in lower case.

Signed-off-by: Pranav Jain <pranavjc@amazon.com>