Fix difference in behaviour for CONCAT() and CONCAT_WS() functions with multibyte characters (#2949)
Issues

Arguments datatype and return datatype of CONCAT() and CONCAT_WS() string function is not defined correctly in
Babelfish, which is causing output difference between TSQL and Babelfish.

CONCAT() and CONCAT_WS() functions does not raise error when called with < 2 or < 3 arguments respectively.

BABEL-4299 missed to add SQL changes to upgrade script, as a result after upgrade to 15.5, 14.10 or later, the
definition will remain same as the previous one and there would be no performance improvement.

Changes made to fix the issues

Added wrapper function handling for CONCAT() and CONCAT_WS() function to ensure correct input arguments
datatype and return datatype. Also ensured to throw appropriate error when CONCAT() and CONCAT_WS() functions
are called with < 2 or < 3 arguments respectively.

Task: BABEL- 3409, BABEL- 3648
Signed-off-by: Anikait Agrawal <agraani@amazon.com>