Added handling to escape unmapped undefined_table and undefined_column error to go to catch block (#3025) (#3035)
Currently, for unmapped errors with PostgreSQL error codes ERRCODE_UNDEFINED_TABLE and ERRCODE_UNDEFINED_COLUMN when thrown in from a lower level of execution in TRY CATCH block it does not go to the CATCH block. This PR adds a handling to fix this issue by allowing unmapped errors with PostgreSQL error code ERRCODE_UNDEFINED_TABLE and ERRCODE_UNDEFINED_COLUMN when thrown from a lower level execution and CATCH block is not inside of trigger, to go to the catch block.

Authored-by: Rohit Bhagat rohitbgt@amazon.com
Signed-off-by: Rohit Bhagat rohitbgt@amazon.com

Issues Resolved
BABEL-5186