Fix dump restore issue for patindex and replace and add MVU test cases for charindex replace patindex (#3020)
Babelfish gives preference to functions with argument of type INT over VARCHAR when trying to select best function when input values for arguments is of unknown type.

Since sys.replace() and sys.patindex() are built in function, we should not rely on this best candidate selection for functions and instead cast the arguments to desired type.

Added MVU test cases for charindex replace and patindex along with these changes

Issues Resolved: https://github.com/babelfish-for-postgresql/babelfish_extensions/issues/2944
                            BABEL-5144

Signed-off-by: Tanzeel Khan <tzlkhan@amazon.com>