Fix regression in FOR JSON AUTO with nested queries. (#3125)
* Fix regression in FOR JSON AUTO with nested queries.

The previous implementation of nested FOR JSON AUTO relied on the happy
accident that the enum T_JsonArrayQueryConstructor happened to have the
same value as the typid for the json type, 114. In PG17 the enum value
for T_JsonArrayQueryConstructor changed to 130, leading to errors when
trying to evaluate nested FOR JSON AUTO queries. Fix this by properly
using the typid for json.

Task: BABEL-5398

Signed-off-by: Jason Teng <jasonten@amazon.com>